optimize something, I don't know

+++ src/main/kotlin/com/fan/db/entity/AnalysisLog.kt
@@ -8,31 +8,22 @@ import jakarta.persistence.Id
 
 @Entity(name = "analysis_log")
 class AnalysisLog(
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    val id: Long? = null,
+    @Id @GeneratedValue(strategy = GenerationType.AUTO) val id: Long? = null,
 
-    @Column(name = "new", nullable = false)
-    val new: Int,
+    @Column(name = "new", nullable = false) val new: Int,
 
-    @Column(name = "valid_titles", nullable = false)
-    val validTitles: Int,
+    @Column(name = "valid_titles", nullable = false) val validTitles: Long,
 
-    @Column(name = "valid_details", nullable = false)
-    val validDetails: Int,
+    @Column(name = "valid_details", nullable = false) val validDetails: Long,
 
-    @Column(name = "success_accounts", nullable = false)
-    val successAccounts: Int,
+    @Column(name = "success_accounts", nullable = false) val successAccounts: Long,
 
-    @Column(name = "failed_accounts", nullable = false)
-    val failedAccounts: Int,
+    @Column(name = "failed_accounts", nullable = false) val failedAccounts: Int,
 
 
-    @Column(name = "type", nullable = false)
-    val type: String,
+    @Column(name = "type", nullable = false) val type: String,
 
-    @Column(name = "date", nullable = false, length = 8196)
-    val date: String,
+    @Column(name = "date", nullable = false, length = 8196) val date: String,
 
     ) {
